chore: extract HijackLogrus to log/hijack package
