+++ log/hijack/logrus.go
@@ -1,4 +1,4 @@
-package envbuilder
+package hijack
 
 import (
 	"io"
@@ -6,10 +6,10 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// HijackLogrus hijacks the logrus logger and calls the callback for each log entry.
+// Logrus hijacks the logrus logger and calls the callback for each log entry.
 // This is an abuse of logrus, the package that Kaniko uses, but it exposes
 // no other way to obtain the log entries.
-func HijackLogrus(callback func(entry *logrus.Entry)) {
+func Logrus(callback func(entry *logrus.Entry)) {
 	logrus.StandardLogger().SetOutput(io.Discard)
 	logrus.StandardLogger().SetFormatter(&logrusFormatter{
 		callback: callback,
